requirements.txt
absl-py==2.1.0
amqp==5.3.1
asgiref==3.8.1
astunparse==1.6.3
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.3.1
cron-descriptor==1.4.5
cycler==0.12.1
Django==5.1.3
django-celery-beat==2.7.0
django-cors-headers==4.6.0
django-rest-framework==0.1.0
django-timezone-field==7.0
djangorestframework==3.15.2
flatbuffers==24.3.25
fonttools==4.55.0
gast==0.6.0
geographiclib==2.0
geopy==2.4.1
google-pasta==0.2.0
grpcio==1.68.0
gunicorn==23.0.0
h5py==3.12.1
idna==3.10
keras==3.6.0
kiwisolver==1.4.7
kombu==5.4.2
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.0.8
numpy==2.0.2
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pandas==2.2.3
pillow==11.0.0
prompt_toolkit==3.0.48
protobuf==5.28.3
psycopg2-binary==2.9.10
Pygments==2.18.0
pygrib==2.1.6
pyparsing==3.2.0
pyproj==3.7.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.2
PyWavelets==1.7.0
redis==5.2.0
requests==2.32.3
rich==13.9.4
setuptools==75.5.0
six==1.16.0
sqlparse==0.5.2
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
termcolor==2.5.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.0
wrapt==1.16.0
